requirements.txt
pytest==8.1.1
pytest-playwright==0.4.4
allure-pytest==2.13.5
